requirements.txt
# progress bars in model download and training scripts
tqdm
# Accessing files from S3 directly.
boto3
# Used for downloading models over HTTP
requests
# For OpenAI GPT
regex != 2019.12.17
# For XLNet
sentencepiece
# For XLM
sacremoses
# For Sentiment3/GLUE:
nltk
ftfy
kafka-python
